refactor: where clauses in dedicated class
