+++ core/Database/Concerns/Query/HasJoinClause.php
@@ -17,9 +17,29 @@ trait HasJoinClause
         return $this;
     }
 
+    public function innerJoinOnEqual(string $relationship, string $column, string $value): self
+    {
+        $join = new Join($relationship, Joins::INNER);
+        $join->onEqual($column, $value);
+
+        $this->pushJoin($join);
+
+        return $this;
+    }
+
     public function leftJoin(string $relationship, Closure $callback): self
     {
         $this->jointIt($relationship, $callback, Joins::LEFT);
+
+        return $this;
+    }
+
+    public function leftJoinOnEqual(string $relationship, string $column, string $value): self
+    {
+        $join = new Join($relationship, Joins::LEFT);
+        $join->onEqual($column, $value);
+
+        $this->pushJoin($join);
 
         return $this;
     }
@@ -58,6 +78,19 @@ trait HasJoinClause
 
         $callback($join);
 
+        $this->pushJoin($join);
+    }
+
+    protected function jointFrom(string $relationship, string $column, string $value): void
+    {
+        $join = new Join($relationship, Joins::INNER);
+        $join->onEqual($column, $value);
+
+        $this->pushJoin($join);
+    }
+
+    protected function pushJoin(Join $join): void
+    {
         [$dml, $arguments] = $join->toSql();
 
         $this->joins[] = $dml;
