+++ core/Database/Query.php
@@ -4,11 +4,10 @@ declare(strict_types=1);
 
 namespace Core\Database;
 
-use BadMethodCallException;
 use Closure;
+use Core\Contracts\Database\Builder;
 use Core\Contracts\Database\QueryBuilder;
 use Core\Database\Concerns\Query\HasJoinClause;
-use Core\Database\Concerns\Query\HasWhereClause;
 use Core\Database\Constants\Actions;
 use Core\Database\Constants\Operators;
 use Core\Database\Constants\Order;
@@ -16,20 +15,14 @@ use Core\Exceptions\QueryError;
 use Core\Util\Arr;
 use Stringable;
 
-class Query implements QueryBuilder
+class Query extends Clause implements QueryBuilder, Builder
 {
     use HasJoinClause;
-    use HasWhereClause;
-
-    public const PLACEHOLDER = '?';
 
     protected readonly string $table;
     protected readonly Actions $action;
     protected array $fields;
-    protected array $where;
     protected array $joins;
-    protected array $arguments;
-    protected Operators|null $logicalConnector;
     protected readonly array $orderBy;
     protected readonly array $limit;
 
@@ -118,50 +111,6 @@ class Query implements QueryBuilder
         ];
     }
 
-    public function __call(string $method, array $arguments = [])
-    {
-        if (str_starts_with($method, 'or')) {
-            $method = lcfirst(str_replace('or', '', $method));
-
-            return $this->setLogicalConnector(Operators::OR)
-                ->{$method}(...$arguments)
-                ->setLogicalConnector(null);
-        }
-
-        if (method_exists($this, $method)) {
-            return $this->{$method}(...$arguments);
-        }
-
-        throw new BadMethodCallException("The method does not exist: {$method}");
-    }
-
-    protected function setLogicalConnector(Operators|null $operator): self
-    {
-        $this->logicalConnector = $operator;
-
-        return $this;
-    }
-
-    protected function pushWhereWithArgs(string $column, Operators $operator, array|string|int $value): void
-    {
-        $placeholders = \is_array($value)
-            ? array_fill(0, count($value), self::PLACEHOLDER)
-            : self::PLACEHOLDER;
-
-        $this->pushWhere([$column, $operator, $placeholders]);
-
-        $this->arguments = array_merge($this->arguments, (array) $value);
-    }
-
-    protected function pushWhere(array $where): void
-    {
-        if (count($this->where) > 0) {
-            array_unshift($where, $this->logicalConnector ?? Operators::AND);
-        }
-
-        $this->where[] = $where;
-    }
-
     protected function buildSelectQuery(): string
     {
         $query = [
@@ -189,34 +138,6 @@ class Query implements QueryBuilder
         return Arr::implodeDeeply($query);
     }
 
-    protected function resolveWhereMethod(string $column, Operators $operator, Closure|array|string|int $value): void
-    {
-        if ($value instanceof Closure) {
-            $this->whereSubquery($value, $operator, $column);
-        } else {
-            $this->pushWhereWithArgs($column, $operator, $value);
-        }
-    }
-
-    protected function whereSubquery(
-        Closure $subquery,
-        Operators $comparisonOperator,
-        string|null $column = null,
-        Operators|null $operator = null
-    ): void {
-        $builder = new Subquery();
-
-        $subquery($builder);
-
-        [$dml, $arguments] = $builder->toSql();
-
-        $value = $operator?->value . $dml;
-
-        $this->pushWhere(array_filter([$column, $comparisonOperator, $value]));
-
-        $this->arguments = array_merge($this->arguments, $arguments);
-    }
-
     protected function prepareFields(array $fields): string
     {
         $fields = array_map(function ($field) {
@@ -228,19 +149,6 @@ class Query implements QueryBuilder
         }, $fields);
 
         return Arr::implodeDeeply($fields, ', ');
-    }
-
-    protected function prepareClauses(array $clauses): array
-    {
-        return array_map(function (array $clause): array {
-            return array_map(function ($value) {
-                return match (true) {
-                    $value instanceof Operators => $value->value,
-                    \is_array($value) => '(' . Arr::implodeDeeply($value, ', ') . ')',
-                    default => $value,
-                };
-            }, $clause);
-        }, $clauses);
     }
 
     private function resolveSubquery(Subquery $subquery): string
