+++ tests/Core/Database/Query/SelectColumnsTest.php
